+++ app/components/UI/TransactionReview/index.js
@@ -1,9 +1,7 @@
 import React, { PureComponent } from 'react';
 import ActionView from '../ActionView';
-import Identicon from '../Identicon';
-import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 import PropTypes from 'prop-types';
-import { StyleSheet, Text, View, PixelRatio, InteractionManager } from 'react-native';
+import { StyleSheet, Text, View, InteractionManager } from 'react-native';
 import { colors, fontStyles } from '../../../styles/common';
 import { connect } from 'react-redux';
 import { strings } from '../../../../locales/i18n';
@@ -13,67 +11,14 @@ import DefaultTabBar from 'react-native-scrollable-tab-view/DefaultTabBar';
 import TransactionReviewInformation from './TransactionReviewInformation';
 import TransactionReviewData from './TransactionReviewData';
 import TransactionReviewSummary from './TransactionReviewSummary';
-import { renderAccountName, safeToChecksumAddress } from '../../../util/address';
 import Analytics from '../../../core/Analytics';
 import ANALYTICS_EVENT_OPTS from '../../../util/analytics';
-import contractMap from 'eth-contract-metadata';
-import AssetIcon from '../AssetIcon';
-import FontAwesome from 'react-native-vector-icons/FontAwesome';
+import TransactionDirection from '../../Views/TransactionDirection';
 
-const FONT_SIZE = PixelRatio.get() < 2 ? 12 : 16;
 const styles = StyleSheet.create({
 	root: {
 		backgroundColor: colors.white,
 		flex: 1
-	},
-	graphic: {
-		borderBottomWidth: 1,
-		borderColor: colors.grey100,
-		borderTopWidth: 1,
-		flexDirection: 'row',
-		flexGrow: 0,
-		flexShrink: 0
-	},
-	addressText: {
-		...fontStyles.bold,
-		fontSize: FONT_SIZE,
-		marginLeft: 9
-	},
-	arrow: {
-		backgroundColor: colors.white,
-		borderColor: colors.grey200,
-		borderRadius: 15,
-		borderWidth: 1,
-		height: 30,
-		width: 30,
-		marginTop: -15,
-		marginLeft: -15,
-		left: '50%',
-		position: 'absolute',
-		zIndex: 1,
-		alignSelf: 'center'
-	},
-	arrowIcon: {
-		color: colors.grey400,
-		marginLeft: 3,
-		marginTop: 3
-	},
-	addressGraphic: {
-		alignItems: 'center',
-		flexDirection: 'row',
-		minHeight: 42,
-		width: '50%',
-		flex: 1
-	},
-	fromGraphic: {
-		borderColor: colors.grey100,
-		borderRightWidth: 1,
-		paddingRight: 35,
-		paddingLeft: 20
-	},
-	toGraphic: {
-		paddingRight: 20,
-		paddingLeft: 35
 	},
 	reviewForm: {
 		flex: 1
@@ -105,22 +50,6 @@ const styles = StyleSheet.create({
 		letterSpacing: 0.5,
 		marginHorizontal: 14,
 		...fontStyles.normal
-	},
-	ensRecipientContainer: {
-		flex: 1,
-		marginLeft: 9
-	},
-	ensRecipient: { ...fontStyles.bold, fontSize: FONT_SIZE },
-	ensAddress: { ...fontStyles.normal, fontSize: 10 },
-	addressWrapper: { flex: 1 },
-	contractLogo: {
-		height: 24,
-		width: 24,
-		backgroundColor: colors.white
-	},
-	contractLogoWrapper: {
-		backgroundColor: colors.white,
-		alignItems: 'flex-start'
 	}
 });
 
@@ -129,10 +58,6 @@ const styles = StyleSheet.create({
  */
 class TransactionReview extends PureComponent {
 	static propTypes = {
-		/**
-		/* Identities object required to get account name
-		*/
-		identities: PropTypes.object,
 		/**
 		 * Callback triggered when this transaction is cancelled
 		 */
@@ -192,77 +117,6 @@ class TransactionReview extends PureComponent {
 		onModeChange && onModeChange('edit');
 	};
 
-	renderToAddressDirection = () => {
-		const {
-			transaction: { to, ensRecipient },
-			identities
-		} = this.props;
-		let child;
-		if (ensRecipient) {
-			child = (
-				<View style={styles.ensRecipientContainer}>
-					<Text style={styles.ensRecipient} numberOfLines={1}>
-						{ensRecipient}
-					</Text>
-					<Text style={styles.ensAddress} numberOfLines={1}>
-						{renderAccountName(to, identities)}
-					</Text>
-				</View>
-			);
-		} else if (to) {
-			child = (
-				<Text style={[styles.addressText, styles.addressWrapper]} numberOfLines={1}>
-					{renderAccountName(to, identities)}
-				</Text>
-			);
-		} else {
-			child = (
-				<Text style={[styles.addressText, styles.addressWrapper]} numberOfLines={1}>
-					{strings('transactions.to_contract')}
-				</Text>
-			);
-		}
-		return (
-			<View style={[styles.addressGraphic, styles.toGraphic]}>
-				{to ? <Identicon address={to} diameter={18} /> : <FontAwesome name="file-text-o" size={18} />}
-				{child}
-			</View>
-		);
-	};
-
-	renderToContractDirection = contract => (
-		<View style={[styles.addressGraphic, styles.toGraphic]}>
-			<View style={styles.contractLogoWrapper}>
-				<AssetIcon logo={contract.logo} customStyle={styles.contractLogo} />
-			</View>
-			<Text style={[styles.addressText, styles.addressWrapper]} numberOfLines={1}>
-				{contract.name}
-			</Text>
-		</View>
-	);
-
-	renderTransactionDirection = () => {
-		const {
-			transaction: { from, to },
-			identities
-		} = this.props;
-		const contract = contractMap[safeToChecksumAddress(to)];
-		return (
-			<View style={styles.graphic}>
-				<View style={[styles.addressGraphic, styles.fromGraphic]}>
-					<Identicon address={from} diameter={18} />
-					<Text style={styles.addressText} numberOfLines={1}>
-						{renderAccountName(from, identities)}
-					</Text>
-				</View>
-				<View style={styles.arrow}>
-					<MaterialIcon name={'arrow-forward'} size={22} style={styles.arrowIcon} />
-				</View>
-				{contract === undefined ? this.renderToAddressDirection() : this.renderToContractDirection(contract)}
-			</View>
-		);
-	};
-
 	renderTabBar() {
 		return (
 			<DefaultTabBar
@@ -301,7 +155,7 @@ class TransactionReview extends PureComponent {
 		const { actionKey, error } = this.state;
 		return (
 			<View style={styles.root}>
-				{this.renderTransactionDirection()}
+				<TransactionDirection />
 				<ActionView
 					confirmButtonMode="confirm"
 					cancelText={strings('transaction.reject')}
@@ -323,7 +177,6 @@ class TransactionReview extends PureComponent {
 
 const mapStateToProps = state => ({
 	accounts: state.engine.backgroundState.AccountTrackerController.accounts,
-	identities: state.engine.backgroundState.PreferencesController.identities,
 	showHexData: state.settings.showHexData,
 	transaction: state.transaction
 });
