+++ scripts/build.sh
@@ -139,7 +139,7 @@ buildAndroid(){
 
 buildIosSimulator(){
 	prebuild_ios
-	react-native run-ios --simulator "iPhone 11 Pro (13.2)"
+	react-native run-ios --simulator "iPhone 11 Pro (13.3)"
 }
 
 buildIosDevice(){
@@ -163,7 +163,7 @@ buildIosRelease(){
 		if [ ! -f "ios/release.xcconfig" ] ; then
 			echo $IOS_ENV | tr "|" "\n" > ios/release.xcconfig
 		fi
-		./node_modules/.bin/react-native run-ios  --configuration Release --simulator "iPhone 11 Pro (13.2)"
+		./node_modules/.bin/react-native run-ios  --configuration Release --simulator "iPhone 11 Pro (13.2.2)"
 	fi
 }
